+++ pkgs/unified_analytics/CHANGELOG.md
@@ -1,7 +1,9 @@
-## 5.5.0-wip
+## 5.5.0
 
 - Edit to the `Event.flutterCommandResult` constructor to add `commandHasTerminal`
 - Added timeout for `Analytics.setTelemetry` to prevent the clients from hanging
+- Added the `Event.appleUsageEvent` constructor
+- Added the `Event.exception` constructor
 
 ## 5.4.0